Refactored Open an account and completed Transfer Money feature works.

+++ src/main/java/com/progotisystemsltd/assigment/mobilebanking/repository/AccountTypeRepository.java
@@ -4,6 +4,6 @@ import com.progotisystemsltd.assigment.mobilebanking.model.AccountType;
 import com.progotisystemsltd.assigment.mobilebanking.model.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountTypeRepository extends JpaRepository<AccountType, Long> {
+public interface AccountTypeRepository extends JpaRepository<AccountType, Integer> {
     AccountType findAccountTypeByAccountTypeName(String accountTypeName);
 }
